requirements.txt
Django
django-cors-headers
djangorestframework
Requests
mysqlclient
psycopg2-binary
Pillow